# contracts/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "CheckoutRequest": {
        "properties": {
          "plan_key": {
            "title": "Plan Key",
            "type": "string"
          }
        },
        "required": [
          "plan_key"
        ],
        "title": "CheckoutRequest",
        "type": "object"
      },
      "CheckoutResponse": {
        "properties": {
          "manual_instructions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Manual Instructions"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "required": [
          "url",
          "provider"
        ],
        "title": "CheckoutResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InvoiceDetailRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "hosted_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hosted Url"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "issued_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issued At"
          },
          "lines": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/InvoiceLineRead"
            },
            "title": "Lines",
            "type": "array"
          },
          "number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Number"
          },
          "paid_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paid At"
          },
          "period_end": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period End"
          },
          "period_start": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period Start"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "subtotal_cents": {
            "title": "Subtotal Cents",
            "type": "integer"
          },
          "tax_cents": {
            "title": "Tax Cents",
            "type": "integer"
          },
          "total_cents": {
            "title": "Total Cents",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "number",
          "currency",
          "subtotal_cents",
          "tax_cents",
          "total_cents",
          "status",
          "period_start",
          "period_end",
          "hosted_url",
          "issued_at",
          "paid_at",
          "created_at"
        ],
        "title": "InvoiceDetailRead",
        "type": "object"
      },
      "InvoiceDraftRequest": {
        "properties": {
          "period": {
            "anyOf": [
              {
                "pattern": "^\\d{4}-\\d{2}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period"
          }
        },
        "title": "InvoiceDraftRequest",
        "type": "object"
      },
      "InvoiceLineRead": {
        "properties": {
          "amount_cents": {
            "title": "Amount Cents",
            "type": "integer"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "metric": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metric"
          },
          "quantity": {
            "title": "Quantity",
            "type": "integer"
          },
          "unit_amount_cents": {
            "title": "Unit Amount Cents",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "kind",
          "description",
          "quantity",
          "unit_amount_cents",
          "amount_cents",
          "metric"
        ],
        "title": "InvoiceLineRead",
        "type": "object"
      },
      "InvoiceRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "hosted_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hosted Url"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "issued_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issued At"
          },
          "number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Number"
          },
          "paid_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paid At"
          },
          "period_end": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period End"
          },
          "period_start": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period Start"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "subtotal_cents": {
            "title": "Subtotal Cents",
            "type": "integer"
          },
          "tax_cents": {
            "title": "Tax Cents",
            "type": "integer"
          },
          "total_cents": {
            "title": "Total Cents",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "number",
          "currency",
          "subtotal_cents",
          "tax_cents",
          "total_cents",
          "status",
          "period_start",
          "period_end",
          "hosted_url",
          "issued_at",
          "paid_at",
          "created_at"
        ],
        "title": "InvoiceRead",
        "type": "object"
      },
...
      "PaymentRecordRequest": {
        "properties": {
          "amount_cents": {
            "exclusiveMinimum": 0.0,
            "title": "Amount Cents",
            "type": "integer"
          },
          "reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference"
          }
        },
        "required": [
          "amount_cents"
        ],
        "title": "PaymentRecordRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Multi-tenant SaaS framework: tenancy, plans, entitlements, usage, billing.",
    "title": "Synapse SaaS Framework",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/billing/admin/invoices/{invoice_id}/pay": {
      "post": {
        "description": "Record an external payment (bank transfer, check, cash) against an open invoice.",
        "operationId": "admin_record_payment_v1_billing_admin_invoices__invoice_id__pay_post",
        "parameters": [
          {
            "in": "path",
            "name": "invoice_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Invoice Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentRecordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceDetailRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Record Payment",
        "tags": [
          "billing"
        ]
      }
    },
    "/v1/billing/admin/invoices/{invoice_id}/void": {
      "post": {
        "operationId": "admin_void_invoice_v1_billing_admin_invoices__invoice_id__void_post",
        "parameters": [
          {
            "in": "path",
            "name": "invoice_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Invoice Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceDetailRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Void Invoice",
        "tags": [
          "billing"
        ]
      }
    },
...
    "/v1/billing/checkout": {
      "post": {
        "operationId": "start_checkout_v1_billing_checkout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Checkout",
        "tags": [
          "billing"
        ]
      }
    },
    "/v1/billing/checkout/confirm": {
      "post": {
        "description": "Offline-payment flow: the tenant confirms, the operator collects out of band.\n\n409 on any provider that verifies payment itself (Stripe, Paddle, Xendit,\nPayMongo) \u2014 those activate from the provider webhook only.",
        "operationId": "confirm_checkout_v1_billing_checkout_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Confirm Checkout V1 Billing Checkout Confirm Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Confirm Checkout",
        "tags": [
          "billing"
        ]
      }
    },
    "/v1/billing/invoices": {
      "get": {
        "operationId": "list_invoices_v1_billing_invoices_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/InvoiceRead"
                  },
                  "title": "Response List Invoices V1 Billing Invoices Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Invoices",
        "tags": [
          "billing"
        ]
      }
    },
    "/v1/billing/invoices/draft": {
      "post": {
        "description": "Generate (or return the existing) draft for the period: plan + overage lines.",
        "operationId": "draft_invoice_v1_billing_invoices_draft_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceDraftRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceDetailRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Draft Invoice",
        "tags": [
          "billing"
        ]
      }
    },
    "/v1/billing/invoices/{invoice_id}": {
      "get": {
        "operationId": "get_invoice_detail_v1_billing_invoices__invoice_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "invoice_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Invoice Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceDetailRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Invoice Detail",
        "tags": [
          "billing"
        ]
      }
    },
    "/v1/billing/invoices/{invoice_id}/finalize": {
      "post": {
        "description": "Assign a number, lock amounts, move to open. Outbox emits for webhooks/email.",
        "operationId": "finalize_invoice_v1_billing_invoices__invoice_id__finalize_post",
        "parameters": [
          {
            "in": "path",
            "name": "invoice_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Invoice Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceDetailRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Finalize Invoice",
        "tags": [
          "billing"
        ]
      }
    },
    "/v1/billing/invoices/{invoice_id}/pdf": {
      "get": {
        "description": "Stream the framework-rendered PDF for an invoice.",
        "operationId": "download_invoice_pdf_v1_billing_invoices__invoice_id__pdf_get",
        "parameters": [
          {
            "in": "path",
            "name": "invoice_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Invoice Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Invoice Pdf",
        "tags": [
          "billing"
        ]
      }
    },
...
    "/v1/billing/webhooks/{provider}": {
      "post": {
        "description": "Provider \u2192 us. Raw body read exactly once before anything parses it.",
        "operationId": "billing_webhook_v1_billing_webhooks__provider__post",
        "parameters": [
          {
            "in": "path",
            "name": "provider",
            "required": true,
            "schema": {
              "title": "Provider",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Billing Webhook V1 Billing Webhooks  Provider  Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Billing Webhook",
        "tags": [
          "billing"
        ]
      }
    },
...
  }
}